CARP OrderedBin: change bin counts to uint64_t everywhere. + IMP BUG FIX in carp_utils (semantics of bins->Size() changed, causing tests to fail)

+++ tests/pivot_calc_10_data.cc
@@ -110,7 +110,7 @@ const float range_min = 0;
 const float range_max = 0;
 const int num_pivots = 2048;
 
-const float rank_bin_counts[] = {
+const uint64_t rank_bin_counts[] = {
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
